+++ packages/netlify-cms-core/src/components/MediaLibrary/MediaLibraryTop.js
@@ -4,7 +4,13 @@ import styled from '@emotion/styled';
 import MediaLibrarySearch from './MediaLibrarySearch';
 import MediaLibraryHeader from './MediaLibraryHeader';
 import MediaLibraryBreadcrumbs from './MediaLibraryBreadCrumbs';
-import { UploadButton, DeleteButton, DownloadButton, InsertButton } from './MediaLibraryButtons';
+import {
+  UploadButton,
+  DeleteButton,
+  DownloadButton,
+  CopyToClipBoardButton,
+  InsertButton,
+} from './MediaLibraryButtons';
 
 const LibraryTop = styled.div`
   position: relative;
@@ -41,12 +47,11 @@ const MediaLibraryTop = ({
   handleBreadcrumbClick,
   currentMediaFolder,
   defaultMediaFolder,
+  selectedFile,
 }) => {
   const shouldShowButtonLoader = isPersisting || isDeleting;
   const uploadEnabled = !shouldShowButtonLoader;
   const deleteEnabled = !shouldShowButtonLoader && hasSelection;
-  const downloadEnabled = hasSelection;
-  const insertEnabled = hasSelection;
 
   const uploadButtonLabel = isPersisting
     ? t('mediaLibrary.mediaLibraryModal.uploading')
@@ -70,11 +75,14 @@ const MediaLibraryTop = ({
           isPrivate={privateUpload}
         />
         <ButtonsContainer>
-          <DownloadButton
-            onClick={onDownload}
-            disabled={!downloadEnabled}
-            focused={downloadEnabled}
-          >
+          <CopyToClipBoardButton
+            disabled={!hasSelection}
+            path={selectedFile.path}
+            name={selectedFile.name}
+            draft={selectedFile.draft}
+            t={t}
+          />
+          <DownloadButton onClick={onDownload} disabled={!hasSelection}>
             {downloadButtonLabel}
           </DownloadButton>
           <UploadButton
@@ -98,7 +106,7 @@ const MediaLibraryTop = ({
             {deleteButtonLabel}
           </DeleteButton>
           {!canInsert ? null : (
-            <InsertButton onClick={onInsert} disabled={!insertEnabled}>
+            <InsertButton onClick={onInsert} disabled={!hasSelection}>
               {insertButtonLabel}
             </InsertButton>
           )}
@@ -132,6 +140,14 @@ MediaLibraryTop.propTypes = {
   hasSelection: PropTypes.bool.isRequired,
   isPersisting: PropTypes.bool,
   isDeleting: PropTypes.bool,
+  selectedFile: PropTypes.oneOfType([
+    PropTypes.shape({
+      path: PropTypes.string.isRequired,
+      draft: PropTypes.bool.isRequired,
+      name: PropTypes.string.isRequired,
+    }),
+    PropTypes.shape({}),
+  ]),
 };
 
 export default MediaLibraryTop;